+++ android/src/test/java/net/glxn/qrgen/android/QRCodeTest.java
@@ -1,162 +1,19 @@
 package net.glxn.qrgen.android;
 
-import android.graphics.Bitmap;
-
 import com.google.zxing.BarcodeFormat;
 import com.google.zxing.EncodeHintType;
 import com.google.zxing.Writer;
 import com.google.zxing.WriterException;
 import com.google.zxing.common.BitMatrix;
 import com.google.zxing.qrcode.decoder.ErrorCorrectionLevel;
-
 import net.glxn.qrgen.core.exception.QRGenerationException;
 import net.glxn.qrgen.core.image.ImageType;
-import net.glxn.qrgen.core.scheme.VCard;
-
 import org.junit.Assert;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.robolectric.RobolectricTestRunner;
-import org.robolectric.annotation.Config;
-import org.robolectric.shadows.ShadowBitmap;
 
-import java.io.ByteArrayOutputStream;
-import java.io.File;
-import java.io.FileOutputStream;
 import java.util.Map;
 
-@RunWith(RobolectricTestRunner.class)
-@Config(manifest = Config.NONE, shadows = {ShadowBitmap.class})
 public class QRCodeTest {
-
-    @Test
-    public void shouldCreateBitmapWithDefaults() throws Exception {
-        Bitmap bmp = QRCode.from("www.example.org").bitmap();
-        Assert.assertNotNull(bmp);
-    }
-
-    @Test
-    public void shouldGetFileAsBitmapWithDefaults() throws Exception {
-        File file = QRCode.from("www.example.org").to(ImageType.BMP).file();
-        Assert.assertNotNull(file);
-    }
-
-    @Test
-    public void shouldGetFileFromVCardWithDefaults() throws Exception {
-        VCard johnDoe = new VCard("John Doe")
-                .setName("John Doe")
-                .setEmail("john.doe@example.org")
-                .setAddress("John Doe Street 1, 5678 Berlin")
-                .setTitle("Mister").setCompany("John Doe Inc.")
-                .setPhoneNumber("1234").setWebsite("www.example.org");
-        File file = QRCode.from(johnDoe).file();
-        Assert.assertNotNull(file);
-    }
-
-    @Test
-    public void shouldGetFileFromTextWithDefaults() throws Exception {
-        File file = QRCode.from("Hello World").file();
-        Assert.assertNotNull(file);
-    }
-
-    @Test
-    public void shouldGetFileWithNameFromTextWithDefaults() throws Exception {
-        File file = QRCode.from("Hello World").file("Hello World");
-        Assert.assertNotNull(file);
-        Assert.assertTrue(file.getName().startsWith("Hello World"));
-    }
-
-    @Test
-    public void shouldGetSTREAMFromTextWithDefaults() throws Exception {
-        ByteArrayOutputStream stream = QRCode.from("Hello World").stream();
-        Assert.assertNotNull(stream);
-    }
-
-    @Test
-    public void shouldHandleLargeString() throws Exception {
-        int length = 2950;
-        char[] chars = new char[length];
-        for (int i = 0; i < length; i++) {
-            chars[i] = 'a';
-        }
-        String text = new String(chars);
-        Assert.assertEquals(length, text.length());
-
-        File file = QRCode.from(text).to(ImageType.PNG).file();
-        Assert.assertNotNull(file);
-    }
-
-    @Test
-    public void shouldGetFileWithNameFromTextWithImageTypeOverrides() throws Exception {
-        File jpg = QRCode.from("Hello World").to(ImageType.JPG)
-                         .file("Hello World");
-        Assert.assertNotNull(jpg);
-        Assert.assertTrue(jpg.getName().startsWith("Hello World"));
-        File gif = QRCode.from("Hello World").to(ImageType.GIF)
-                         .file("Hello World");
-        Assert.assertNotNull(gif);
-        Assert.assertTrue(gif.getName().startsWith("Hello World"));
-    }
-
-    @Test
-    public void shouldGetStreamFromText() throws Exception {
-        ByteArrayOutputStream stream = QRCode.from("Hello World").to(ImageType.PNG).stream();
-        Assert.assertNotNull(stream);
-        File tempFile = File.createTempFile("test", ".tmp");
-        long lengthBefore = tempFile.length();
-        FileOutputStream fileOutputStream = new FileOutputStream(tempFile);
-        stream.writeTo(fileOutputStream);
-        Assert.assertTrue(lengthBefore < tempFile.length());
-    }
-
-    @Test
-    public void shouldWriteToSuppliedStream() throws Exception {
-        ByteArrayOutputStream stream = new ByteArrayOutputStream();
-        QRCode.from("Hello World").writeTo(stream);
-
-        Assert.assertNotNull(stream);
-        File tempFile = File.createTempFile("test", ".tmp");
-        long lengthBefore = tempFile.length();
-        FileOutputStream fileOutputStream = new FileOutputStream(tempFile);
-        stream.writeTo(fileOutputStream);
-        Assert.assertTrue(lengthBefore < tempFile.length());
-    }
-
-    @Test
-    public void shouldBeAbleToOverrideDimensionsToFile() throws Exception {
-        long defaultSize = QRCode.from("Hello World").to(ImageType.PNG).file().length();
-        long defaultSize2 = QRCode.from("Hello World").to(ImageType.PNG).file().length();
-        File file = QRCode.from("Hello World").to(ImageType.PNG).withSize(250, 250).file();
-        Assert.assertNotNull(file);
-        Assert.assertTrue(defaultSize == defaultSize2);
-        //This seems not to work due to some limitations of robolectric, the genrated file on disk is also broken
-        //Assert.assertTrue(defaultSize < file.length());
-    }
-
-    @Test
-    public void shouldBeAbleToOverrideDimensionsToFileWithName() throws Exception {
-        long defaultSize = QRCode.from("Hello World").to(ImageType.PNG).file("Hello World").length();
-        long defaultSize2 = QRCode.from("Hello World").to(ImageType.PNG).file("Hello World").length();
-        File file = QRCode.from("Hello World").to(ImageType.PNG).withSize(250, 250).file("Hello World");
-        Assert.assertNotNull(file);
-        Assert.assertTrue(defaultSize == defaultSize2);
-        //This seems not to work due to some limitations of robolectric, the genrated file on disk is also broken
-        //Assert.assertTrue(defaultSize < file.length());
-        Assert.assertTrue(file.getName().startsWith("Hello World"));
-    }
-
-    @Test
-    public void shouldBeAbleToSupplyEncodingHint() throws Exception {
-        String expected = "UTF-8";
-        final Object[] capture = new Object[1];
-        try {
-            final QRCode from = QRCode.from("Jour férié");
-            from.setQrWriter(writerWithCapture(capture));
-            from.to(ImageType.PNG).withCharset(expected).stream();
-        } catch (QRGenerationException ignored) {
-        }
-        assertCapturedHint(expected, capture, EncodeHintType.CHARACTER_SET);
-    }
 
     @Test
     public void shouldBeAbleToSupplyErrorCorrectionHint() throws Exception {
@@ -166,7 +23,7 @@ public class QRCodeTest {
             final QRCode from = QRCode.from("Jour férié");
             from.setQrWriter(writerWithCapture(capture));
             from.to(ImageType.PNG).withErrorCorrection(ErrorCorrectionLevel.L)
-                .stream();
+                    .stream();
         } catch (QRGenerationException ignored) {
         }
         assertCapturedHint(expected, capture, EncodeHintType.ERROR_CORRECTION);
@@ -186,6 +43,19 @@ public class QRCodeTest {
             }
             assertCapturedHint(expected, capture, type);
         }
+    }
+
+    @Test
+    public void shouldBeAbleToSupplyEncodingHint() throws Exception {
+        String expected = "UTF-8";
+        final Object[] capture = new Object[1];
+        try {
+            final QRCode from = QRCode.from("Jour férié");
+            from.setQrWriter(writerWithCapture(capture));
+            from.to(ImageType.PNG).withCharset(expected).stream();
+        } catch (QRGenerationException ignored) {
+        }
+        assertCapturedHint(expected, capture, EncodeHintType.CHARACTER_SET);
     }
 
     @SuppressWarnings("unchecked")
