+++ pom.xml
@@ -42,29 +42,12 @@
         <module>android</module>
     </modules>
 
-    <properties>
-        <zxing.version>3.4.1</zxing.version>
-        <batik.version>1.14</batik.version>
-        <junit.version>4.8.2</junit.version>
-        <robolectric.version>3.0</robolectric.version>
-        <fest-android.version>1.0.7</fest-android.version>
-        <android.platform.version>4.1.1.4</android.platform.version>
-        <maven-compiler-plugin.source>1.7</maven-compiler-plugin.source>
-        <maven-compiler-plugin.target>1.7</maven-compiler-plugin.target>
-        <maven-compiler-plugin.version>2.3.2</maven-compiler-plugin.version>
-        <maven-checkstyle-plugin.version>2.12.1</maven-checkstyle-plugin.version>
-        <maven-javadoc-plugin.version>3.0.1</maven-javadoc-plugin.version>
-        <maven-source-plugin.version>3.0.1</maven-source-plugin.version>
-        <nexus-staging-maven-plugin.version>1.6.4</nexus-staging-maven-plugin.version>
-        <maven-gpg-plugin.version>1.5</maven-gpg-plugin.version>
-    </properties>
-
     <reporting>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-checkstyle-plugin</artifactId>
-                <version>${maven-checkstyle-plugin.version}</version>
+                <version>2.12.1</version>
                 <configuration>
                     <configLocation>checkstyle.xml</configLocation>
                 </configuration>
@@ -79,43 +62,39 @@
         </plugins>
     </reporting>
 
-    <profiles>
-        <profile>
-            <id>sonatype-oss-release</id>
-            <distributionManagement>
-              <repository>
-                <id>github</id>
-                <name>GitHub OWNER Apache Maven Packages</name>
-                <url>https://maven.pkg.github.com/OWNER/REPOSITORY</url>
-              </repository>
-            </distributionManagement>
-        </profile>
-    </profiles>
-
     <build>
-        <plugins>
-            <plugin>
-                <artifactId>maven-compiler-plugin</artifactId>
-                <version>${maven-compiler-plugin.version}</version>
-                <configuration>
-                    <source>${maven-compiler-plugin.source}</source>
-                    <target>${maven-compiler-plugin.target}</target>
-                    <encoding>UTF-8</encoding>
-                </configuration>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-source-plugin</artifactId>
-                <version>${maven-source-plugin.version}</version>
-                <executions>
-                    <execution>
-                        <goals>
-                            <goal>jar</goal>
-                        </goals>
-                    </execution>
-                </executions>
-            </plugin>
-        </plugins>
+        <pluginManagement>
+            <plugins>
+                <plugin>
+                    <artifactId>maven-compiler-plugin</artifactId>
+                    <version>3.8.0</version>
+                    <configuration>
+                        <release>11</release>
+                        <encoding>UTF-8</encoding>
+                    </configuration>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-source-plugin</artifactId>
+                    <version>3.0.1</version>
+                    <executions>
+                        <execution>
+                            <goals>
+                                <goal>jar</goal>
+                            </goals>
+                        </execution>
+                    </executions>
+                </plugin>
+                <plugin>
+                    <groupId>org.apache.maven.plugins</groupId>
+                    <artifactId>maven-javadoc-plugin</artifactId>
+                    <version>3.0.1</version>
+                    <configuration>
+                        <javadocExecutable>${java.home}/bin/javadoc</javadocExecutable>
+                    </configuration>
+                </plugin>
+            </plugins>
+        </pluginManagement>
     </build>
 
     <dependencyManagement>
@@ -123,22 +102,22 @@
             <dependency>
                 <groupId>com.google.zxing</groupId>
                 <artifactId>javase</artifactId>
-                <version>${zxing.version}</version>
+                <version>3.4.1</version>
             </dependency>
             <dependency>
                 <groupId>com.google.zxing</groupId>
                 <artifactId>core</artifactId>
-                <version>${zxing.version}</version>
+                <version>3.4.1</version>
             </dependency>
             <dependency>
                 <groupId>org.apache.xmlgraphics</groupId>
                 <artifactId>batik-svggen</artifactId>
-                <version>${batik.version}</version>
+                <version>1.14</version>
             </dependency>
             <dependency>
                 <groupId>org.apache.xmlgraphics</groupId>
                 <artifactId>batik-dom</artifactId>
-                <version>${batik.version}</version>
+                <version>1.14</version>
             </dependency>
             <dependency>
                 <groupId>org.assertj</groupId>
@@ -149,25 +128,13 @@
             <dependency>
                 <groupId>com.google.android</groupId>
                 <artifactId>android</artifactId>
-                <version>${android.platform.version}</version>
+                <version>4.1.1.4</version>
                 <scope>provided</scope>
             </dependency>
             <dependency>
                 <groupId>junit</groupId>
                 <artifactId>junit</artifactId>
-                <version>${junit.version}</version>
-                <scope>test</scope>
-            </dependency>
-            <dependency>
-                <groupId>org.robolectric</groupId>
-                <artifactId>robolectric</artifactId>
-                <version>${robolectric.version}</version>
-                <scope>test</scope>
-            </dependency>
-            <dependency>
-                <groupId>com.squareup</groupId>
-                <artifactId>fest-android</artifactId>
-                <version>${fest-android.version}</version>
+                <version>4.13.2</version>
                 <scope>test</scope>
             </dependency>
         </dependencies>
